// src/models/mietpreisData.ts
export interface MietpreisData {
    _id: number;
    StichtagDatJahr: string;
    StichtagDatMonat: string;
    RaumeinheitSort: string;
    RaumeinheitLang: string;
    GliederungSort: string;
    GliederungLang: string;
    ZimmerSort: string;
    ZimmerLang: string;
    GemeinnuetzigSort: string;
    GemeinnuetzigLang: string;
    EinheitSort: string;
    EinheitLang: string;
    PreisartSort: string;
    PreisartLang: string;
    mean: string;
    meanl: string;
    meanu: string;
    qu10: string;
    qu10l: string;
    qu10u: string;
    qu25: string;
    qu25l: string;
    qu25u: string;
    qu50: string;
    qu50l: string;
    qu50u: string;
    qu75: string;
    qu75l: string;
    qu75u: string;
    qu90: string;
    qu90l: string;
    qu90u: string;
    Domain: string;
    Sample1: string;
    Sample2: string;
}